python_fastapi/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
fastapi[standard]
uvicorn[standard] # install httptools and uvloop
orjson
sqlmodel
pydantic[email]
pydantic-settings
aiosqlite
httpx
authlib
asgi-lifespan
python-multipart
pytest
pytest-asyncio
pytest-mock
pytest-cov